requirements.txt
eth-brownie
eth_abi
web3
websockets
